+++ src/main/java/ru/mail/polis/service/alexander/marashov/ServiceImpl.java
@@ -39,6 +39,8 @@ import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 
+import static ru.mail.polis.service.alexander.marashov.ValidatedParameters.validateParameters;
+
 public class ServiceImpl extends HttpServer implements Service {
 
     private static final Logger log = LoggerFactory.getLogger(ServiceImpl.class);
@@ -131,28 +133,6 @@ public class ServiceImpl extends HttpServer implements Service {
         } catch (final IOException ioException) {
             log.error(RESPONSE_ERROR_STRING, ioException);
         }
-    }
-
-    private boolean areParametersWrong(
-            final String id,
-            final int ack,
-            final int from
-    ) {
-        return id.isEmpty() || ack <= 0 || ack > from || from > topology.size();
-    }
-
-    private static int[] unpackReplicasParameter(final String replicas) throws NumberFormatException {
-        if (replicas == null) {
-            return new int[0];
-        }
-        final List<String> parameters = Splitter.on('/').splitToList(replicas);
-        if (parameters.size() != 2) {
-            return new int[0];
-        }
-
-        final int ack = Integer.parseInt(parameters.get(0));
-        final int from = Integer.parseInt(parameters.get(1));
-        return new int[]{ack, from};
     }
 
     @Override
@@ -249,9 +229,9 @@ public class ServiceImpl extends HttpServer implements Service {
         executeOrSendError(
                 httpSession,
                 () -> {
-                    final ValidatedParameters validatedParameters;
+                    final ValidatedParameters validParams;
                     try {
-                        validatedParameters = validateParameters(id, replicas);
+                        validParams = validateParameters(id, replicas, defaultAck, defaultFrom, topology.size());
                     } catch (final IllegalArgumentException e) {
                         sendAnswerOrError(httpSession, new Response(Response.BAD_REQUEST, Response.EMPTY));
                         return;
@@ -259,13 +239,13 @@ public class ServiceImpl extends HttpServer implements Service {
                     final String rowAccessHeader = request.getHeader(ROW_ACCESS_HEADER);
 
                     if (rowAccessHeader != null && rowAccessHeader.equals(Integer.toString(Request.METHOD_GET))) {
-                        sendAnswerOrError(httpSession, executeLocalRowGet(validatedParameters.key));
+                        sendAnswerOrError(httpSession, executeLocalRowGet(validParams.key));
                         return;
                     }
 
-                    final String[] primaries = topology.primariesFor(validatedParameters.key, validatedParameters.from);
+                    final String[] primaries = topology.primariesFor(validParams.key, validParams.from);
                     final ResponseAnalyzerGet valueAnalyzer =
-                            new ResponseAnalyzerGet(validatedParameters.ack, validatedParameters.from);
+                            new ResponseAnalyzerGet(validParams.ack, validParams.from);
 
                     request.addHeader(ROW_ACCESS_HEADER + Request.METHOD_GET);
                     for (final String primary : primaries) {
@@ -274,7 +254,7 @@ public class ServiceImpl extends HttpServer implements Service {
                             executorService.execute(() -> {
                                 Value value = null;
                                 try {
-                                    value = this.dao.rowGet(validatedParameters.key);
+                                    value = this.dao.rowGet(validParams.key);
                                 } catch (final NoSuchElementException | IOException e) {
                                     log.debug("Key not found", e);
                                 }
@@ -328,9 +308,9 @@ public class ServiceImpl extends HttpServer implements Service {
         executeOrSendError(
                 httpSession,
                 () -> {
-                    final ValidatedParameters validatedParameters;
+                    final ValidatedParameters validParams;
                     try {
-                        validatedParameters = validateParameters(id, replicas);
+                        validParams = validateParameters(id, replicas, defaultAck, defaultFrom, topology.size());
                     } catch (final IllegalArgumentException e) {
                         sendAnswerOrError(httpSession, new Response(Response.BAD_REQUEST, Response.EMPTY));
                         return;
@@ -341,14 +321,14 @@ public class ServiceImpl extends HttpServer implements Service {
 
                     final String rowAccessHeader = request.getHeader(ROW_ACCESS_HEADER);
                     if (rowAccessHeader != null && rowAccessHeader.equals(Integer.toString(Request.METHOD_PUT))) {
-                        sendAnswerOrError(httpSession, executeLocalPut(validatedParameters.key, value));
+                        sendAnswerOrError(httpSession, executeLocalPut(validParams.key, value));
                         return;
                     }
 
-                    final String[] primaries = topology.primariesFor(validatedParameters.key, validatedParameters.from);
+                    final String[] primaries = topology.primariesFor(validParams.key, validParams.from);
                     final ResponseAnalyzer<Boolean> responseAnalyzer = new SimpleResponseAnalyzer(
-                            validatedParameters.ack,
-                            validatedParameters.from,
+                            validParams.ack,
+                            validParams.from,
                             201,
                             Response.CREATED
                     );
@@ -357,7 +337,7 @@ public class ServiceImpl extends HttpServer implements Service {
                     for (final String primary : primaries) {
                         if (topology.isLocal(primary)) {
                             executorService.execute(() -> {
-                                final Response response = executeLocalPut(validatedParameters.key, value);
+                                final Response response = executeLocalPut(validParams.key, value);
                                 responseAnalyzer.accept(response);
                             });
                         } else {
@@ -385,50 +365,6 @@ public class ServiceImpl extends HttpServer implements Service {
         );
     }
 
-    private static class ValidatedParameters {
-        final int ack;
-        final int from;
-        final ByteBuffer key;
-
-        public ValidatedParameters(final int ack, final int from, final ByteBuffer key) {
-            this.ack = ack;
-            this.from = from;
-            this.key = key;
-        }
-    }
-
-    /**
-     * Validates parameters and throws exceptions if parameters are invalid.
-     * @param id - id parameter.
-     * @param replicas - replicas parameter, has ack/from format.
-     * @throws IllegalArgumentException if any parameter has invalid value.
-     * @return validated parameters.
-     */
-    private ValidatedParameters validateParameters(
-            final String id,
-            final String replicas
-    ) throws IllegalArgumentException {
-        final int[] replicasParameters = unpackReplicasParameter(replicas);
-        final int ack;
-        final int from;
-        if (replicasParameters.length == 0) {
-            ack = defaultAck;
-            from = defaultFrom;
-        } else {
-            ack = replicasParameters[0];
-            from = replicasParameters[1];
-        }
-
-        if (areParametersWrong(id, ack, from)) {
-            throw new IllegalArgumentException("Invalid arguments");
-        }
-
-        final byte[] bytes = id.getBytes(StandardCharsets.UTF_8);
-        final ByteBuffer key = ByteBuffer.wrap(bytes);
-
-        return new ValidatedParameters(ack, from, key);
-    }
-
     /**
      * HTTP method handler for removing a value by the key from the DAO storage.
      *
@@ -449,9 +385,9 @@ public class ServiceImpl extends HttpServer implements Service {
         executeOrSendError(
                 httpSession,
                 () -> {
-                    final ValidatedParameters validatedParameters;
+                    final ValidatedParameters validParams;
                     try {
-                        validatedParameters = validateParameters(id, replicas);
+                        validParams = validateParameters(id, replicas, defaultAck, defaultFrom, topology.size());
                     } catch (final IllegalArgumentException e) {
                         sendAnswerOrError(httpSession, new Response(Response.BAD_REQUEST, Response.EMPTY));
                         return;
@@ -459,14 +395,14 @@ public class ServiceImpl extends HttpServer implements Service {
 
                     final String rowAccessHeader = request.getHeader(ROW_ACCESS_HEADER);
                     if (rowAccessHeader != null && rowAccessHeader.equals(Integer.toString(Request.METHOD_DELETE))) {
-                        sendAnswerOrError(httpSession, executeLocalDelete(validatedParameters.key));
+                        sendAnswerOrError(httpSession, executeLocalDelete(validParams.key));
                         return;
                     }
 
-                    final String[] primaries = topology.primariesFor(validatedParameters.key, validatedParameters.from);
+                    final String[] primaries = topology.primariesFor(validParams.key, validParams.from);
                     final ResponseAnalyzer<Boolean> responseAnalyzer = new SimpleResponseAnalyzer(
-                            validatedParameters.ack,
-                            validatedParameters.from,
+                            validParams.ack,
+                            validParams.from,
                             202,
                             Response.ACCEPTED
                     );
@@ -475,7 +411,7 @@ public class ServiceImpl extends HttpServer implements Service {
                     for (final String primary : primaries) {
                         if (topology.isLocal(primary)) {
                             executorService.execute(() -> {
-                                final Response response = executeLocalDelete(validatedParameters.key);
+                                final Response response = executeLocalDelete(validParams.key);
                                 responseAnalyzer.accept(response);
                             });
                         } else {
